requirements.txt
# ============================================================================
# Health AI Chatbot - Streamlit Cloud Requirements
# Optimized for cloud deployment (lightweight, no heavy ML frameworks)
# ============================================================================

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Visualization
plotly>=5.17.0

# AI/LLM Integration - Gemini API
google-generativeai==0.3.2

# Lightweight RAG Components (No LangChain, No PyTorch, No Transformers)
qdrant-client==1.7.0
sentence-transformers==2.3.1
faiss-cpu
PyPDF2==3.0.1
torch>=2.2.0  # torch==2.1.0 not available for this env

# Phase 2: Advanced RAG improvements
rank-bm25  # Sparse keyword retrieval
transformers  # For cross-encoder re-ranking
